requirements.txt
tinker-api>=0.1.0
datasets>=2.14.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
sympy>=1.12
scipy>=1.11.0
